socketcan: fix socketcan-raw added
