+++ Cargo.toml
@@ -36,6 +36,7 @@ mio = "1"
 serial_test = "2.0"
 smol = "1.3"
 socket2 = "0.5"
+socketcan-raw = { path = "raw" }
 thiserror = "2"
 tokio = "1"
 
@@ -55,9 +56,9 @@ description.workspace = true
 
 # Features:
 #
-# "netlink" (default) - Whether to include CAN interface configuration 
+# "netlink" (default) - Whether to include CAN interface configuration
 #       capabilities based on netlink kernel communications
-# "dump" (default) - Whether to include 'candump' log file parsing 
+# "dump" (default) - Whether to include 'candump' log file parsing
 #	capabilities.
 # "enumerate" - Ability to enumerate the available CAN network interfaces
 #
@@ -79,7 +80,7 @@ utils = ["clap", "anyhow"]
 tokio = ["dep:tokio", "mio", "futures"]
 async-std = ["dep:async-std", "dep:async-io"]
 async-io = ["dep:async-io"]
-enumerate = ["dep:libudev"]
+enumerate = ["dep:libudev", "socketcan-raw/enumerate"]
 
 [dependencies]
 embedded-can = { workspace = true }
@@ -101,6 +102,7 @@ async-io = { workspace = true, optional = true }
 smol = { workspace = true, optional = true }
 async-std = { workspace = true, optional = true }
 libudev = { workspace = true, optional = true }
+socketcan-raw = { workspace = true, features = ["embedded-can"] }
 
 # This hack avoids building neli on non-linux platforms, which
 # avoids a ton of compile errors. Worthwhile together with the
