+++ examples/smol_send.rs
@@ -17,7 +17,8 @@
 
 use embedded_can::{Frame, StandardId};
 use futures_timer::Delay;
-use socketcan::{smol::CanSocket, CanFrame, Error, Result};
+use socketcan::{smol::CanSocket, CanFrame};
+use socketcan_raw::{Error, Result};
 use std::{env, time::Duration};
 
 fn main() -> Result<()> {
