+++ src/lib.rs
@@ -125,8 +125,6 @@
     unsafe_op_in_unsafe_fn
 )]
 
-use std::mem::size_of;
-
 // Re-export the embedded_can crate so that applications can rely on
 // finding the same version we use.
 pub use embedded_can::{
@@ -135,13 +133,6 @@ pub use embedded_can::{
 };
 
 pub mod errors;
-pub use errors::{
-    CanError, CanErrorDecodingFailure, ConstructionError, Error, IoError, IoErrorKind, IoResult,
-    Result,
-};
-
-pub mod addr;
-pub use addr::CanAddr;
 
 pub mod id;
 pub use id::CanId;
@@ -155,8 +146,8 @@ pub use frame::{
 #[cfg(feature = "dump")]
 pub mod dump;
 
-pub mod socket;
-pub use socket::{CanFdSocket, CanFilter, CanSocket, ShouldRetry, Socket, SocketOptions};
+mod socket;
+pub use socket::{can::CanSocket, can_fd::CanFdSocket};
 
 #[cfg(feature = "netlink")]
 pub mod nl;
@@ -188,20 +179,3 @@ pub mod async_std {
 pub mod enumerate;
 #[cfg(feature = "enumerate")]
 pub use enumerate::available_interfaces;
-
-// ===== helper functions =====
-
-/// Gets a byte slice for any sized variable.
-///
-/// Note that this should normally be unsafe, but since we're only
-/// using it internally for types sent to the kernel, it's OK.
-pub(crate) fn as_bytes<T: Sized>(val: &T) -> &[u8] {
-    let sz = size_of::<T>();
-    unsafe { std::slice::from_raw_parts::<'_, u8>(val as *const _ as *const u8, sz) }
-}
-
-/// Gets a mutable byte slice for any sized variable.
-pub(crate) fn as_bytes_mut<T: Sized>(val: &mut T) -> &mut [u8] {
-    let sz = size_of::<T>();
-    unsafe { std::slice::from_raw_parts_mut(val as *mut _ as *mut u8, sz) }
-}
