+++ src/socket/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod can;
+pub(crate) mod can_fd;
